feat: margin shorthand property
Your marker would look better if it was centered on the page. An easy way to do that is with the margin shorthand property.

In the last project, you set the 'margin' area of elements separately with properties like 'margin-top' and 'margin-left'. The 'margin' shorthand property makes it easy to set multiple margin areas at the same time.

To center your marker on the page, set its 'margin' property to 'auto'. This sets 'margin-top', 'margin-right', 'margin-bottom', and 'margin-left' all to 'auto'.