feat: rgb color model
Earlier you learned that the RGB color model is additive. This means that colors begin as black, and change as different levels of red, green, and blue are introduced.

An easy way to see this is with the CSS 'rgb' function.

Create a new CSS rule that targets the class 'container' and set its 'background-color' to black with 'rgb(0, 0, 0)'.