feat: shorthand padding
Now add a little more vertical space between your markers and the edge of the container element they're in.

In the '.container 'CSS rule, use the shorthand 'padding' property to add '10px' of top and bottom padding, and set the left and right padding to '0'. This works similarly to the shorthand 'margin' property you used earlier.