feat: primary rgb max value-white
In the additive RGB color model, primary colors are colors that, when combined, create pure white. But for this to happen, each color needs to be at its highest intensity.

Now that you have the primary RGB colors, it's time to combine them.

For the 'rgb 'function in the '.container' rule, set the red, green, and blue values to the max of '255'.