feat: color wheel
A color wheel is a circle where similar colors, or hues, are near each other, and different ones are further apart. For example, pure red is between the hues rose and orange.

Two colors that are opposite from each other on the color wheel are called complementary colors. If two complementary colors are combined, they produce gray. But when they are placed side-by-side, these colors produce strong visual contrast and appear brighter.

In the 'rgb' function for the '.one' CSS rule, set the red value to the max of '255' to produce pure red. In the 'rgb' function for '.two' CSS rule, set the values for green and blue to the max of '255' to produce cyan.