feat: color gradient
You've learned a few ways to set flat colors in CSS, but you can also use a color transition, or gradient, on an element.

A gradient is when one color transitions into another. The CSS 'linear-gradient' function lets you control the direction of the transition along a line, and which colors are used.

One thing to remember is that the 'linear-gradient' function actually creates an 'image' element, and is usually paired with the 'background' property which can accept an image as a value.

In the '.red' CSS rule, change the 'background-color' property to 'background'.