feat: border-left shorthand property
The border-left shorthand property lets you to set the left border's width, style, and color at the same time.

Here is the syntax:

border-left: width style color;

In the '.sleeve' CSS rule, replace the 'border-left-width', 'border-left-style', and 'border-left-color' properties with the 'border-left' shorthand property. The values for the width, style, and color of the left border should be the same.